ui margin으로 수정, 라우터 연결
